+++ src/test/issues.c2.test.ts
@@ -2014,4 +2014,12 @@ describe("Issue Tests", () => {
         equal(returnTypes, expectedTypes);
         equal(paramTypes, expectedTypes);
     });
+
+    it("#2842 handles computed properties with @class", () => {
+        const project = convert();
+        const hello = query(project, "ComputedClass.hello");
+
+        equal(hello.kind, ReflectionKind.Property);
+        equal(hello.type?.toString(), "string");
+    });
 });